Fixed up new callback system and changes, after I retested
Broke up different gui things in new files, as now they do not have to be glomped in together

Debugged changes in the innermost layer (again)

Now I can add the new gui stuff
