+++ index.html
@@ -129,7 +129,7 @@
 
 <!-- My Stuff-->
 <script src="js/errors.js"></script>
-<script src="js/gui/helpers.js"></script>
+<script src="js/gui/toast.js"></script>
 
 
 <script src="js/shell/shell-game-serialized.js"></script>
@@ -148,8 +148,15 @@
 <script src="js/keeper/keeper.js"></script>
 
 
-<script src="js/gui/editor_gui.js"></script>
-<script src="js/gui/gui_init.js"></script>
+
 <script src="js/gui/game-main.js"></script>
+<script src="js/gui/editor_gui.js"></script>
+
+<script src="js/gui/gui-select-shell.js"></script>
+<script src="js/gui/gui-tags.js"></script>
+<script src="js/gui/load.js"></script>
+<script src="js/gui/step.js"></script>
+<script src="js/gui/toast.js"></script>
+<script src="js/gui/extras.js"></script>
 </body>
 </html>
